Получение категорий и лотов из базы данных. Распредение функций по файлам.
