+++ config.php
@@ -1,0 +1,10 @@
+<?php
+return [
+    'db' =>
+        [
+            'host' => 'localhost',
+            'username' => 'root',
+            'password' => '',
+            'database' => 'yeticave'
+        ]
+];
