+++ functions/template.php
@@ -1,0 +1,18 @@
+<?php
+
+function include_template($name, $data) {
+    $name = 'templates/' . $name;
+    $result = '';
+
+    if (!file_exists($name)) {
+        return $result;
+    }
+
+    ob_start();
+    extract($data);
+    require $name;
+
+    $result = ob_get_clean();
+
+    return $result;
+}
