+++ templates/index.php
@@ -4,7 +4,7 @@
 			<ul class="promo__list">
 			  <?php foreach ($categories as $category) : ?>
 				<li class="promo__item promo__item--boards">
-					<a class="promo__link" href="pages/all-lots.html"><?= $category ?></a>
+					<a class="promo__link" href="pages/all-lots.html"><?= $category['name'] ?></a>
 				</li>
 			  <?php endforeach; ?>
 			</ul>
@@ -17,15 +17,15 @@
 			  <?php foreach ($lots as $key => $lot) : ?>
 				<li class="lots__item lot">
 					<div class="lot__image">
-						<img src="<?= htmlspecialchars($lot['picture']) ?>" width="350" height="260" alt="">
+						<img src="<?= htmlspecialchars($lot['image']) ?>" width="350" height="260" alt="">
 					</div>
 					<div class="lot__info">
-						<span class="lot__category"><?= htmlspecialchars($lot['category']) ?></span>
-						<h3 class="lot__title"><a class="text-link" href="pages/lot.html"><?= $lot['name'] ?></a></h3>
+						<span class="lot__category"><?= htmlspecialchars($lot['category_name']) ?></span>
+						<h3 class="lot__title"><a class="text-link" href="pages/lot.html"><?= $lot['title'] ?></a></h3>
 							<div class="lot__state">
 								<div class="lot__rate">
 									<span class="lot__amount">Стартовая цена</span>
-										<span class="lot__cost"><?= cut_num($lot['price']) ?></span>
+										<span class="lot__cost"><?= format_price($lot['price']) ?></span>
 								</div>
 								<div class="lot__timer timer">
                                     <?= time_before_tomorrow() ?>
